+++ components/restreamer/client/cypress/integration/add_output_mixin.js
@@ -39,8 +39,8 @@ describe('ADD MIXIN OUTPUT', () => {
   });
 
   it('Submits', () => {
-    cy.get("button:contains('Add')").click();
-    cy.get("button:contains('Add')").should('not.exist');
+    cy.get('button').contains(/^Add$/).click();
+    cy.get('button').contains(/^Add$/).should('not.exist');
   });
 
   it('Assert', () => {