+++ components/restreamer/client/cypress/integration/add_output_multiple.js
@@ -20,8 +20,8 @@ describe('ADD MULTIPLE OUTPUT', () => {
   });
 
   it('Submits', () => {
-    cy.get("button:contains('Add')").click();
-    cy.get("button:contains('Add')").should('not.exist');
+    cy.get('button').contains(/^Add$/).click();
+    cy.get('button').contains(/^Add$/).should('not.exist');
   });
 
   it('Assert', () => {