+++ components/restreamer/client/src/modals/RestreamModal.svelte
@@ -1,71 +1,85 @@
-<script lang="js">
+<script lang="ts">
   import { onDestroy } from 'svelte';
-  import { get } from 'svelte/store';
   import { mutation } from 'svelte-apollo';
-  import { SetRestream } from '../../api/client.graphql';
-  import { showError } from '../utils/util';
-  import { restreamModal as value } from '../stores';
+  import { SetRestream } from '../../api/client.js';
+  import { sanitizeLabel, showError } from '../utils/util';
   import { saveOrCloseByKeys } from '../utils/directives.util';
+  import { RestreamModel } from '../models/restream.model';
+  import { writable } from 'svelte/store';
+  import cloneDeep from 'lodash/cloneDeep';
+  import isEqual from 'lodash/isEqual';
+  import RestreamBackup from './RestreamBackup.svelte';
 
   const setRestreamMutation = mutation(SetRestream);
 
+  export let visible = false;
   export let public_host = 'localhost';
+
+  export let restream: RestreamModel = new RestreamModel();
+  let previous: RestreamModel = cloneDeep(restream);
+
+  let restreamStore = writable(restream);
 
   let submitable = false;
   onDestroy(
-    value.subscribe((v) => {
-      submitable = v.key !== '';
-      let changed = !v.edit_id;
-      if (!!v.edit_id) {
-        changed |=
-          v.key !== v.prev_key ||
-          v.label !== v.prev_label ||
-          v.is_pull !== v.prev_is_pull ||
-          v.with_backup !== v.prev_with_backup;
+    restreamStore.subscribe((current) => {
+      submitable = current.key !== '';
+      let changed = !current.id;
+
+      if (!!current.id) {
+        changed ||=
+          current.key !== previous.key ||
+          current.label !== previous.label ||
+          current.isPull !== previous.isPull;
       }
-      if (v.is_pull) {
-        submitable &= v.pull_url !== '';
-        if (!!v.edit_id) {
-          changed |= v.pull_url !== v.prev_pull_url;
+
+      if (current.isPull) {
+        submitable &&= current.pullUrl !== '';
+        if (!!current.id) {
+          changed ||= current.pullUrl !== previous.pullUrl;
         }
       }
-      if (v.with_backup) {
-        if (!!v.edit_id) {
-          changed |= v.backup_is_pull !== v.prev_backup_is_pull;
-        }
-        if (v.backup_is_pull) {
-          submitable &= v.backup_pull_url !== '';
-          if (!!v.edit_id) {
-            changed |= v.backup_pull_url !== v.prev_backup_pull_url;
-          }
-        }
+
+      if (current.backups.length !== previous.backups.length) {
+        changed ||= true;
+      } else {
+        current.backups.forEach((x, i) => {
+          changed ||= !isEqual(x, previous.backups[i]);
+        });
       }
-      if (!!v.edit_id) {
-        changed |= v.with_hls !== v.prev_with_hls;
+
+      if (!!current.id) {
+        changed ||= current.withHls !== previous.withHls;
       }
-      submitable &= changed;
+      submitable &&= changed;
     })
   );
 
-  async function submit() {
+  async function submit(): Promise<void> {
     if (!submitable) return;
-    const v = get(value);
 
-    let variables = { key: v.key, with_hls: v.with_hls, with_backup: false };
-    if (v.label !== '') {
-      variables.label = v.label;
+    let variables: any = {
+      key: restream.key,
+      with_hls: restream.withHls,
+    };
+
+    if (restream.label) {
+      variables.label = restream.label;
     }
-    if (v.is_pull) {
-      variables.url = v.pull_url;
+
+    if (restream.isPull) {
+      variables.url = restream.pullUrl;
     }
-    if (v.with_backup) {
-      variables.with_backup = true;
-      if (v.backup_is_pull) {
-        variables.backup_url = v.backup_pull_url;
-      }
+
+    if (restream.backups.length) {
+      variables.backup_inputs = restream.backups.map((x) => ({
+        key: x.key,
+        src: x.pullUrl,
+      }));
     }
-    if (v.edit_id) {
-      variables.id = v.edit_id;
+
+    if (restream.id) {
+      variables.id = restream.id;
     }
 
     try {
@@ -76,124 +90,148 @@
     }
   }
 
-  function close() {
-    value.close();
-  }
+  const close = () => {
+    visible = false;
+  };
+
+  const removeBackup = (index: number) => {
+    restreamStore.update((v) => {
+      v.removeBackup(index);
+      return v;
+    });
+  };
+
+  const addBackup = () => {
+    restreamStore.update((v) => {
+      v.addBackup();
+      return v;
+    });
+  };
+
+  const onChangeLabel = () => {
+    restreamStore.update((v) => {
+      v.label = sanitizeLabel(v.label);
+      return v;
+    });
+  };
+
+  const onChangeRestreamKey = () => {
+    restreamStore.update((v) => {
+      v.key = sanitizeLabel(v.key);
+      return v;
+    });
+  };
+
+  const onChangeBackup = () => {
+    restreamStore.update((v) => {
+      return v;
+    });
+  };
 </script>
 
 <template>
-  {#if $value.visible}
-    <div
-      class="uk-modal uk-open"
-      use:saveOrCloseByKeys={{ save: submit, close: close }}
-    >
-      <div class="uk-modal-dialog uk-modal-body">
-        <h2 class="uk-modal-title">
-          {#if $value.edit_id}Edit{:else}Add new{/if} input source for re-streaming
-        </h2>
-        <button
-          class="uk-modal-close-outside"
-          uk-close
-          type="button"
-          on:click={close}
-        />
+  <div
+    class="uk-modal uk-open"
+    use:saveOrCloseByKeys={{ save: submit, close: close }}
+  >
+    <div class="uk-modal-dialog uk-modal-body">
+      <h2 class="uk-modal-title">
+        {#if $restreamStore.id}Edit{:else}Add new{/if} input source for re-streaming
+      </h2>
+      <button
+        class="uk-modal-close-outside"
+        uk-close
+        type="button"
+        on:click={close}
+      />
 
-        <fieldset>
-          <div class="restream">
-            <input
-              class="uk-input uk-form-small"
+      <fieldset>
+        <div class="restream">
+          <input
+            class="uk-input uk-form-small"
+            type="text"
+            data-testid="add-input-modal:label-input"
+            bind:value={$restreamStore.label}
+            on:change={onChangeLabel}
+            placeholder="optional label"
+          />
+          <label
+            >rtmp://{public_host}/<input
+              class="uk-input"
               type="text"
-              data-testid="add-input-modal:label-input"
-              bind:value={$value.label}
-              on:change={() => value.sanitizeLabel()}
-              placeholder="optional label"
+              data-testid="add-input-modal:stream-key-input"
+              placeholder="<stream-key>"
+              bind:value={$restreamStore.key}
+              on:change={onChangeRestreamKey}
+            />/origin</label
+          >
+          <div class="uk-alert">
+            {#if $restreamStore.isPull}
+              Server will pull RTMP stream from the address below.
+              <br />
+              Supported protocols:
+              <code>rtmp://</code>,
+              <code>http://.m3u8</code> (HLS)
+            {:else}
+              Server will await RTMP stream to be pushed onto the address above.
+            {/if}
+          </div>
+        </div>
+        <div class="pull">
+          <label
+            ><input
+              class="uk-checkbox"
+              type="checkbox"
+              bind:checked={$restreamStore.isPull}
+            /> or pull from</label
+          >
+          {#if $restreamStore.isPull}
+            <input
+              class="uk-input"
+              type="text"
+              bind:value={$restreamStore.pullUrl}
+              placeholder="rtmp://..."
             />
-            <label
-              >rtmp://{public_host}/<input
-                class="uk-input"
-                type="text"
-                data-testid="add-input-modal:stream-key-input"
-                placeholder="<stream-key>"
-                bind:value={$value.key}
-              />/origin</label
-            >
-            <div class="uk-alert">
-              {#if $value.is_pull}
-                Server will pull RTMP stream from the address below.
-                <br />
-                Supported protocols:
-                <code>rtmp://</code>,
-                <code>http://.m3u8</code> (HLS)
-              {:else}
-                Server will await RTMP stream to be pushed onto the address
-                above.
-              {/if}
-            </div>
-          </div>
-          <div class="pull">
-            <label
-              ><input
-                class="uk-checkbox"
-                type="checkbox"
-                bind:checked={$value.is_pull}
-              /> or pull from</label
-            >
-            {#if $value.is_pull}
-              <input
-                class="uk-input"
-                type="text"
-                bind:value={$value.pull_url}
-                placeholder="rtmp://..."
-              />
-            {/if}
-          </div>
-          <div class="backup">
-            <label
-              ><input
-                class="uk-checkbox"
-                type="checkbox"
-                bind:checked={$value.with_backup}
-              /> with backup</label
-            >
-            {#if $value.with_backup}
-              <label
-                ><input
-                  class="uk-checkbox"
-                  type="checkbox"
-                  bind:checked={$value.backup_is_pull}
-                /> pulled from</label
-              >
-              {#if $value.backup_is_pull}
-                <input
-                  class="uk-input"
-                  type="text"
-                  bind:value={$value.backup_pull_url}
-                  placeholder="rtmp://..."
-                />
-              {/if}
-            {/if}
-          </div>
-          <div class="hls">
-            <label
-              ><input
-                class="uk-checkbox"
-                type="checkbox"
-                bind:checked={$value.with_hls}
-              /> with HLS endpoint</label
-            >
-          </div>
-        </fieldset>
+          {/if}
+        </div>
+        <div class="hls">
+          <label
+            ><input
+              class="uk-checkbox"
+              type="checkbox"
+              bind:checked={$restreamStore.withHls}
+            /> with HLS endpoint</label
+          >
+        </div>
 
-        <button
-          class="uk-button uk-button-primary"
-          data-testid="add-input-modal:confirm"
-          disabled={!submitable}
-          on:click={submit}
-          >{#if $value.edit_id}Edit{:else}Add{/if}</button
-        >
-      </div>
+        <div class="uk-section uk-section-xsmall backups-section">
+          <button
+            class="uk-button uk-button-primary uk-button-small"
+            on:click={() => addBackup()}
+            >Add backup
+          </button>
+          <ul class="uk-list uk-margin-left">
+            {#each $restreamStore.backups as backup, index}
+              <RestreamBackup
+                {backup}
+                removeFn={() => removeBackup(index)}
+                onChangeFn={() => onChangeBackup()}
+              />
+            {/each}
+          </ul>
+        </div>
+      </fieldset>
+
+      <button
+        class="uk-button uk-button-primary"
+        data-testid="add-input-modal:confirm"
+        disabled={!submitable}
+        on:click={submit}
+      >
+        {#if $restreamStore.id}Edit{:else}Add{/if}
+      </button>
     </div>
-  {/if}
+  </div>
 </template>
 
 <style lang="stylus">
@@ -230,7 +268,8 @@
     .uk-input
       margin-bottom: 10px
 
-  .backup
-    label + label
-      margin-left: 15px
+  .backups-section
+    padding-top: 10px;
+    padding-bottom: 0;
+
 </style>